refactor: aggregated all the Configuration constants
This change rearranges constants present in various Options classes into a single class.
